Issue-9633 - Port ArrayExcept to functions-array subcrate
