+++ datafusion/proto/proto/datafusion.proto
@@ -662,7 +662,7 @@ enum ScalarFunction {
   // 120 was ArrayUnion
   OverLay = 121;
   // 122 is Range
-  ArrayExcept = 123;
+  // 123 is ArrayExcept
   // 124 was ArrayPopFront
   Levenshtein = 125;
   SubstrIndex = 126;
